+++ client/src/lib/mui/theme.js
@@ -33,4 +33,64 @@ const theme = createTheme({
   },
 })
 
-export default theme
+const themeOptionsLight = {
+  palette: {
+    mode: 'light',
+    primary: {
+      main: '#e6d5b3',
+    },
+    secondary: {
+      main: '#d6a3e0',
+    },
+    error: {
+      main: '#d32f2f',
+    },
+    warning: {
+      main: '#ed6c02',
+    },
+    info: {
+      main: '#0288d1',
+    },
+    success: {
+      main: '#2e7d32',
+    },
+    typography: {
+      fontFamily: roboto.style.fontFamily,
+    },
+  },
+}
+
+const themeOptionsDark = {
+  palette: {
+    mode: 'dark',
+    primary: {
+      main: '#e0c2a3',
+    },
+    secondary: {
+      main: '#d6a3e0',
+    },
+    error: {
+      main: '#e0a3a3',
+    },
+    warning: {
+      main: '#e0bea3',
+    },
+    info: {
+      main: '#a3c9e0',
+    },
+    success: {
+      main: '#a3e0aa',
+    },
+    typography: {
+      fontFamily: roboto.style.fontFamily,
+    },
+  },
+}
+const lightTheme = createTheme(themeOptionsLight)
+const darkTheme = createTheme(themeOptionsDark)
+
+export {
+  theme,
+  lightTheme,
+  darkTheme
+}
